nimble/host: Change 'move_count' to uint8_t type to prevent out-of-bounds errors

+++ nimble/host/store/config/src/ble_store_config.c
@@ -181,7 +181,7 @@ ble_store_config_delete_obj(void *values, int value_size, int idx,
 {
     uint8_t *dst;
     uint8_t *src;
-    int move_count;
+    uint8_t move_count;
 
     (*num_values)--;
     if (idx < *num_values) {
